ConsolidateData class which creates, stores and exports MBAR analysis pandas DataFrame.
The completed workflow results will be stored in '.cache/' directory

+++ implicit_solvent_ddm/postTreatment.py
@@ -25,6 +25,189 @@ AVAGADRO = 6.0221367e23
 BOLTZMAN = 1.380658e-23
 JOULES_PER_KCAL = 4184
 
+
+class ConsolidateData(Job):
+    
+    def __init__(self, complex_adative_run, 
+                 ligand_adaptive_run, receptor_adaptive_run,
+                 temperature:float, max_conformation_force,
+                 max_orientational_force,
+                 boresch_df: RestraintMaker,
+                 working_path, complex_filename, 
+                 ligand_filename, receptor_filename, 
+                 ):
+        Job.__init__(self,  memory="2G", cores=2, disk="3G")
+        self.temp = temperature
+        self.complex_adative_run = complex_adative_run
+        self.receptor_adaptive_run = receptor_adaptive_run 
+        self.ligand_adaptive_run = ligand_adaptive_run
+        
+        self.max_con_force = str(max_conformation_force)
+        self.max_orien_force = str(max_orientational_force)
+        self.boresch = boresch_df
+        self.complex_name = re.sub(r"\..*", "", os.path.basename(complex_filename))
+        self.ligand_name = re.sub(r"\..*", "", os.path.basename(ligand_filename))
+        self.receptor_name = re.sub(r"\..*", "", os.path.basename(receptor_filename))
+        self.working_path = working_path
+        #self._set_kcal_per_mol()
+        
+    # def _set_kcal_per_mol(self):
+        
+    #     self.complex_fe_kcal_per_mol = self.complex_fe * self.kcals_per_Kt
+    #     self.complex_error_kcal_per_mol = self.complex_error * self.kcals_per_Kt
+    #     self.ligand_fe_kcal_per_mol  = self.ligand_fe * self.kcals_per_Kt
+    #     self.ligand_error_kcal_per_mol  = self.ligand_error * self.kcals_per_Kt
+    #     self.receptor_fe_kcal_per_mol  = self.receptor_fe * self.kcals_per_Kt
+    #     self.receptor_error_kcal_per_mol  = self.receptor_error * self.kcals_per_Kt
+    @property
+    def complex_mbar_formatted_df(self)->pd.DataFrame:
+        
+        return self.complex_adative_run[1] * self.kcals_per_Kt
+    
+    @property 
+    def complex_fe(self)->pd.DataFrame:
+        
+        return self.complex_adative_run[0][0] * self.kcals_per_Kt
+    
+    @property
+    def complex_error(self)->pd.DataFrame:
+        
+        return self.complex_adative_run[0][1] * self.kcals_per_Kt
+    
+    @property
+    def ligand_mbar_formatted_df(self)->pd.DataFrame:
+        
+        return self.ligand_adaptive_run[1] * self.kcals_per_Kt
+    
+    @property
+    def ligand_fe(self)->pd.DataFrame:
+        
+        return self.ligand_adaptive_run[0][0] * self.kcals_per_Kt
+    
+    @property
+    def ligand_error(self)->pd.DataFrame:
+    
+        return self.ligand_adaptive_run[0][1] * self.kcals_per_Kt
+    
+    @property
+    def receptor_mbar_formatted_df(self)->pd.DataFrame:
+        
+        return self.receptor_adaptive_run[1] * self.kcals_per_Kt
+    
+    
+    @property
+    def receptor_fe(self):
+        
+        return self.receptor_adaptive_run[0][0] * self.kcals_per_Kt
+    
+    @property
+    def receptor_error(self)->pd.DataFrame:
+        
+        return self.receptor_adaptive_run[0][1] * self.kcals_per_Kt
+    
+    
+    @property
+    def kcals_per_Kt(self):
+        
+        return ((BOLTZMAN * (AVAGADRO)) / JOULES_PER_KCAL) * self.temp
+    
+
+    @property
+    def _get_ligand_deltaG(self):
+        
+        return self.ligand_fe.loc[("endstate", "78.5", "1.0", "0.0"), [("electrostatics", "0.0", "0.0", self.max_con_force)]].values[0]
+    
+    @property
+    def _get_receptor_deltaG(self):
+        
+        return self.receptor_fe.loc[("endstate", "78.5", "1.0", "0.0"), [("no_gb", "0.0", "1.0", self.max_con_force)]].values[0]
+    
+    @property
+    def _get_complex_deltaG(self):
+        
+        return self.complex_fe.loc[("no_interactions", "0.0", "0.0", f"{self.max_con_force}_{self.max_orien_force}"), [("endstate", "78.5", "1.0", "0.0_0.0")]].values[0]
+    
+    @property
+    def _get_boresch_standard_state(self):
+        
+        return self.boresch.boresch_deltaG["DeltaG"].values[0]
+   
+    @property
+    def compute_binding_deltaG(self)->float:
+
+        return self._get_complex_deltaG + self._get_ligand_deltaG + self._get_receptor_deltaG + self._get_boresch_standard_state
+    
+    
+    def run(self, fileStore):
+        
+        output_path = os.path.join(f"{self.working_path}", f".cache/{self.complex_name}")    
+        
+        if not os.path.exists(output_path):
+            os.makedirs(output_path)
+        # parse out formatted dataframe
+        self.complex_mbar_formatted_df.to_hdf(
+            f"{output_path}/{self.complex_name}_formatted.h5", key="df", mode="w"
+        )
+        self.receptor_mbar_formatted_df.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_formatted.h5", key="df", mode="w"
+        )
+        self.ligand_mbar_formatted_df.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_formatted.h5", key="df", mode="w"
+        )
+        
+        # parse out free energies
+        self.complex_fe.to_hdf(
+        f"{output_path}/{self.complex_name}_fe.h5", key="df", mode="w"
+        )
+        self.receptor_fe.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_fe.h5", key="df", mode="w"
+        )
+        self.ligand_fe.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_fe.h5", key="df", mode="w"
+        )
+        
+        
+        # parse out errors of mbar
+        self.complex_error.to_hdf(
+            f"{output_path}/{self.complex_name}_error.h5", key="df", mode="w"
+        )
+        self.receptor_error.to_hdf(
+            f"{output_path}/receptor_{self.receptor_name}_error.h5", key="df", mode="w"
+        )
+        self.ligand_error.to_hdf(
+            f"{output_path}/ligand_{self.ligand_name}_error.h5", key="df", mode="w"
+        )
+        
+        # parse out boresch restraints dataframe
+
+        self.boresch.boresch_deltaG.to_hdf(
+            f"{output_path}/boresch_{self.complex_name}.h5", key="df", mode="w"
+        )
+        deltaG_df = pd.DataFrame()
+        
+        
+        fileStore.logToMaster(f"BORESCH standard state {self._get_boresch_standard_state}\n")
+        
+        fileStore.logToMaster(f"Ligand Delta G: {self._get_ligand_deltaG}")
+        fileStore.logToMaster(f"Receptor Delta G: {self._get_receptor_deltaG}\n")
+        
+        fileStore.logToMaster(f"Complex unique index keys:\n {self.complex_fe.index.unique()}\n")
+         
+        fileStore.logToMaster(f"Complex Delta G: {self._get_complex_deltaG}\n")
+
+        deltaG_df[f"{self.ligand_name}_endstate->no_charges"] = [self._get_ligand_deltaG]
+        deltaG_df[f"{self.receptor_name}_endstate->no_gb"] = [self._get_receptor_deltaG]
+        deltaG_df["boresch_restraints"] = [self._get_boresch_standard_state]
+        deltaG_df[f"{self.complex_name}_no-interactions->endstate"] = [
+           self._get_complex_deltaG
+        ]
+        deltaG_df["deltaG"] = [self.compute_binding_deltaG]
+
+        deltaG_df.to_hdf(
+            f"{output_path}/deltaG_{self.complex_name}.h5", key="df", mode="w"
+        )
+        
+        
 class PostTreatment(Job):
     def __init__(
         self,
@@ -202,6 +385,7 @@ def create_mdout_dataframe(
 
     mdout = f"{output_dir}/mdout"
 
+    job.log(f"List files in postprocess directory {os.listdir(output_dir)}\n")
     run_args = sim.dirStruct.fromPath2Dict(mdout)
     data = min_to_dataframe(mdout)
 
@@ -231,6 +415,7 @@ def create_mdout_dataframe(
         data.to_parquet(f"{output_dir}/simulation_mdout.parquet.gzip", compression='gzip') 
         #data.to_parquet(f"{output_dir}/simulation_mdout.zip",  compression="gzip")
     
-    os.remove(mdout)
+    if os.path.exists(mdout):
+        os.remove(mdout)
     
     return data
